test: update search page component tests after refactor
