+++ frontend/src/components/search/filters/FilterGroup.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import FilterGroup from '@/components/search/filters/FilterGroup';
-import { activitiesOptions } from '@/components/search/test/mock-data.test';
+import { activitiesOptions } from '@/components/search/test/mock-data';
 
 vi.mock('react-router-dom', async () => {
   return {
