+++ frontend/src/components/search/filters/FilterMenu.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import FilterMenu from '@/components/search/filters/FilterMenu';
 import searchResultsStore from '@/store/searchResults';
-import { mockFilterMenuContent } from '@/components/search/test/mock-data.test';
+import { mockFilterMenuContent } from '@/components/search/test/mock-data';
 
 Object.defineProperty(searchResultsStore, 'state', {
   get: vi.fn(() => ({
